Revert "Merge remote-tracking branch 'origin/master' into Menu"
This reverts commit 4cf5cc74e6fcf7f9878fdb064fc58a23c18a09e7.

+++ Assets/Scripts/Controllers/DamageableHitboxController.cs
@@ -27,7 +27,7 @@ namespace Controllers
             FindOwnerHealthController();
         }
 
-        public void Hit(float damage)
+        public void Hit(int damage)
         {
             Debug.Log("GUNHIT");
             _healthController?.TakeDamage(damage);
