+++ Assets/Scripts/Strategy/IBullet.cs
@@ -11,6 +11,6 @@ namespace Strategy
         void Travel();
         
         void SetRange(float range);
-        void SetDamage(float damage);
+        void SetDamage(int damage);
     }
 }
